Review and remove deprecated store files
