+++ great_expectations/data_context/store/__init__.py
@@ -1,25 +1,10 @@
-# from .store import (
-#     Store,
-# )
-
-# from .basic import (
-#     FilesystemStore,
-#     # InMemoryStore,
-# )
-
-# from .namespaced import (
-#     # NamespacedStore,
-#     # NamespacedInMemoryStore,
-#     # NamespacedFilesystemStore,
-# )
-
 from .store_backend import (
     StoreBackend,
     InMemoryStoreBackend,
     FilesystemStoreBackend,
 )
 
-from .new_store import (
+from .store import (
     WriteOnlyStore,
     ReadWriteStore,
     BasicInMemoryStore,
